refactor(slice-45): move preflight modules into server.core.guards
Carve the Behavior-theme guards package out of flat server/core so theme
separation can land incrementally; keep one-release shim re-exports and
retarget coverage/patch paths to the canonical package.
