+++ docs/plan/DECISIONS.md
@@ -146,3 +146,8 @@
 | 142 | 2026-07-27 | coverage | Shared product floors **95/90/95/95** for FE; BE **95/90/n/a/95** (functions n/a on coverage.py) | FE Vitest-enforced. BE: gap tests (`test_coverage_floor_gaps.py`) + `fail_under` **90→95** + `scripts/check_backend_coverage_floors.py` (stmts/br/lines from JSON; floors in `[tool.rag_params_finder.coverage_thresholds]`). Wired in quality-gates / pre-push / CI. Measured stmts≈98.6 / br≈95.2 / TOTAL≈97.7. Supersedes #141 soft BE 92/85 | Keep BE fail_under=90 / aim-only without JSON checker |
 | 140 | 2026-07-27 | coverage | Uniform product coverage floor **≥90%** for backend and frontend | Backend unit scoped gate `--cov-fail-under` / `fail_under` **80→90** (measured TOTAL ≈93.2%). Frontend Vitest thresholds briefly **90/90/90/90**. Postgres `retriever_postgres` CI module gate stays **≥95%** (specialized). **FE quartet superseded by #141**; BE fail_under=90 retained until #142 | Keep asymmetric FE 95 / BE 80 |
 | 139 | 2026-07-27 | 44 | Phase B reopen: FE coverage gate → **≥95%** stmts/funcs/lines + **≥90%** branches (`all: true` on `src/**`; exclude tests/setup/main/types) | User chose option 1 + reopen 44 (not 44B). Surpasses backend `--cov-fail-under=80`. Supersedes #132/#138 “100% Won’t / 64% floor” for product thresholds. **Superseded for product floors by #140** | Keep 64% floor / invent literal 100% without exclusions / new Slice 44B |
+| 143 | 2026-07-27 | health-check | Gap 8: Specification coverage gates inserted into SLICE-45 After-Checks | Active PLANNED stub lacked §12 coverage / branch / mutation After-Check lines; user confirmed (y); product floors #142 retained (not whole-tree 100%) | SKIPPED — leave Gap 8 open on Slice 45 |
+| 144 | 2026-07-27 | continuation | Path A Resume — Slice **45** (not first TRAIL 🔨 32) | User redirected resume to module-theme-separation after Gap 8; 32/32C/32B remain open Must chain; 45 is non-blocking Could hygiene | Resume 32 only / leave 45 PLANNED |
+| 145 | 2026-07-27 | review | nw-solution-architect-reviewer Slice 45 roadmap iter2 → **APPROVED** (0 blocking) | Post–Gap 8 + resume; #137 remediations still hold; Before-Checks live count **1/10** (orchestrator corrected reviewer’s 8/11 claim) | NEEDS_REVISION / re-open Must move tables |
+| 146 | 2026-07-27 | 45 | Phase 1 skateboard = `core/guards/` then `pipeline/` orchestrator SLAP (not whole `core/` in one PR) | Blast-radius core ≈112 match-lines (&lt;200); CLI touches `search_index_*`; highest cognitive load still orchestrator — sequence guards first for safer import rewrite | Move entire `server/core/` in one PR / FE-primitives-first |
+| 147 | 2026-07-27 | 45 | Landed `server/core/guards/` + shim re-exports at old `server.core.*` paths; test patches retargeted to `server.core.guards.*` | Preserve caller imports for one release; mypy-safe explicit re-exports; coverage cov paths → guards | sys.modules alias shims / rewrite all callers in same PR |
