+++ docs/plan/TRAIL.md
@@ -37,7 +37,7 @@ Each PCTO / migration slice lives in its own file below. Existing planned slices
 | 38 | [../plan/slices/SLICE-38-CUTOVER-ADR-004.md](../plan/slices/SLICE-38-CUTOVER-ADR-004.md) | Quality comparison + ADR-004 + default cutover | Must | ✅ COMPLETE | 37 | [#118](https://github.com/neomatrix369/rag-params-finder/pull/118) | ~2 min | 2026-07-26 |
 | 43 | [../plan/slices/SLICE-43-SUPABASE-CONFIG-VERIFICATION.md](../plan/slices/SLICE-43-SUPABASE-CONFIG-VERIFICATION.md) | Supabase example-config verification + operator QoL | Could | ✅ COMPLETE | 35 (soft: 37) | [#115](https://github.com/neomatrix369/rag-params-finder/pull/115) | ~2 min | 2026-07-26 |
 | 44 | [../plan/slices/SLICE-44-FRONTEND-COVERAGE-GATE.md](../plan/slices/SLICE-44-FRONTEND-COVERAGE-GATE.md) | Frontend coverage + gate summary + structure taxonomy Should (theme map; moves → 45); review remediations #137 | Should | ✅ COMPLETE | none (Vitest harness) | — | ~3 min | 2026-07-27 |
-| 45 | [../plan/slices/SLICE-45-MODULE-THEME-SEPARATION.md](../plan/slices/SLICE-45-MODULE-THEME-SEPARATION.md) | Module theme separation — execute Behavior/Feature/Function folder moves; architect APPROVED post-#137 | Could | 📋 PLANNED | 44 (taxonomy artifacts) | — | ~2 min | 2026-07-27 |
+| 45 | [../plan/slices/SLICE-45-MODULE-THEME-SEPARATION.md](../plan/slices/SLICE-45-MODULE-THEME-SEPARATION.md) | Module theme separation — execute Behavior/Feature/Function folder moves; architect APPROVED post-#137 | Could | 🔨 IN PROGRESS | 44 (taxonomy artifacts) | — | ~2 min | 2026-07-27 |
 | 28 | [../plan/slices/SLICE-28-RESULTS-EXPORT.md](../plan/slices/SLICE-28-RESULTS-EXPORT.md) | Results export — CSV/JSONL download (issue #49; @cschanhniem) | Must | 📋 PLANNED | none | [#49](https://github.com/neomatrix369/rag-params-finder/issues/49) | ~3 min | 2026-07-06 |
 | 22 | [../plan/slices/SLICE-22-SIE-SCOOTER.md](../plan/slices/SLICE-22-SIE-SCOOTER.md) | SIE Scooter — reranking + SPLADE v3 sparse + `/api/v1/best-config` | Must | 📋 PLANNED | 21, 32, 38 (soft) | — | ~3 min | 2026-07-09 |
 | 26 | [../plan/slices/SLICE-26-LOCAL-MONGODB-DOCS.md](../plan/slices/SLICE-26-LOCAL-MONGODB-DOCS.md) | Local MongoDB: smooth path docs + script feedback | Should | 📦 DEFERRED | 25B | — | ~1.5 min | 2026-07-09 |
