+++ server/core/guards/__init__.py
@@ -1,0 +1,5 @@
+"""Preflight / health guards — Behavior theme (Slice 45).
+
+Import from submodules (e.g. ``server.core.guards.search_index_plan``).
+Temporary shims remain at ``server.core.<module>`` for one release.
+"""
