+++ tests/test_coverage_floor_gaps.py
@@ -120,7 +120,7 @@ def test_given_postgres_extension_row_when_present_then_true() -> None:
     Slice: 44 — BE coverage floor parity
     """
     # -- Given --
-    with patch("server.core.search_index_guard.fetch_one", return_value={"ok": 1}):
+    with patch("server.core.guards.search_index_guard.fetch_one", return_value={"ok": 1}):
         # -- When / Then --
         assert postgres_vector_extension_present() is True
 
@@ -146,7 +146,7 @@ def test_given_required_indexes_when_collect_postgres_snapshot_then_present_set(
     # -- Given --
     required = frozenset({"chunks_embedding_384_hnsw"})
     with patch(
-        "server.core.search_index_guard.fetch_all",
+        "server.core.guards.search_index_guard.fetch_all",
         return_value=[{"indexname": "chunks_embedding_384_hnsw"}],
     ):
         # -- When --
