+++ tests/test_sweep_endpoint.py
@@ -140,8 +140,8 @@ class TestHealthEnhanced:
     def test_check_sie_health_returns_string(self):
         """check_sie_health returns a string status."""
         with (
-            patch("server.core.sie_guard.settings") as mock_settings,
-            patch("server.core.sie_guard.httpx.get") as mock_get,
+            patch("server.core.guards.sie_guard.settings") as mock_settings,
+            patch("server.core.guards.sie_guard.httpx.get") as mock_get,
         ):
             mock_settings.sie_enabled = True
             mock_settings.sie_endpoint = "http://localhost:8720"
@@ -155,7 +155,7 @@ class TestHealthEnhanced:
 
     def test_check_sie_health_returns_disabled_when_sie_off(self):
         """check_sie_health returns 'disabled' when SIE_ENABLED is false."""
-        with patch("server.core.sie_guard.settings") as mock_settings:
+        with patch("server.core.guards.sie_guard.settings") as mock_settings:
             mock_settings.sie_enabled = False
             from server.core.sie_guard import check_sie_health
 
@@ -165,8 +165,11 @@ class TestHealthEnhanced:
     def test_check_sie_health_unreachable_on_exception(self):
         """check_sie_health returns 'unreachable' when SIE is down."""
         with (
-            patch("server.core.sie_guard.settings") as mock_settings,
-            patch("server.core.sie_guard.httpx.get", side_effect=Exception("connection refused")),
+            patch("server.core.guards.sie_guard.settings") as mock_settings,
+            patch(
+                "server.core.guards.sie_guard.httpx.get",
+                side_effect=Exception("connection refused"),
+            ),
         ):
             mock_settings.sie_enabled = True
             mock_settings.sie_endpoint = "http://localhost:8720"
@@ -196,8 +199,8 @@ class TestHealthEnhanced:
 
         test_client = TestClient(app)
         with (
-            patch("server.core.sie_guard.settings") as mock_settings,
-            patch("server.core.sie_guard.httpx.get") as mock_get,
+            patch("server.core.guards.sie_guard.settings") as mock_settings,
+            patch("server.core.guards.sie_guard.httpx.get") as mock_get,
         ):
             mock_settings.sie_enabled = True
             mock_settings.sie_endpoint = "http://localhost:8720"
